Moved form related components to submission page
