+++ src/pages/OpenProblems/OpenProblems.jsx
@@ -11,23 +11,27 @@ import config from "../../utils/configs/SideNavConfig";
 function OpenProblems() {
   const modalState = useSelector((state) => state.question.modalOpen);
   const questionDetails = useSelector((state) => state.question.modalDetails);
+  const filterOpen = useSelector((state) => state.question.filterOpen);
   const dispatch = useDispatch();
   const modalCloseHandler = () => {
     dispatch(generalActions.toggleModal({ bool: false }));
   };
 
-  // Set openProblems
-
   return (
-    <div className="w-full flex flex-row p-6  space-x-4">
-      <div className="side-nav w-1/5 sticky">
+    <div className="w-full flex flex-row space-x-4">
+      {/* Filter Side Navigation */}
+      <div className={`side-nav w-1/5 sticky`}>
         <Filter config={config} />
       </div>
-      <div className="main w-4/5 h-max overflow-auto">
+
+      {/* Main Content */}
+      <div className="main w-full h-max overflow-auto">
         <SearchBar label={"Search for an open problem"} />
         <Statbar className="statbar" />
         <ProblemsInterface />
       </div>
+
+      {/* Modal */}
       <Modal open={modalState} close={modalCloseHandler}>
         <SubmissionModalContent
           questionDetails={questionDetails}
